% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{addmtDNA}
\alias{addmtDNA}
\title{add mtDNA haplotypes to a completed simulation}
\usage{
addmtDNA(indiv = makeFounders(), mtDNAfreqs)
}
\arguments{
\item{indiv}{Individual matrix, as from \code{makeFounders()}. Every
animal in indiv must either be a founder (see
\code{makeFounders()}), or be descended exclusively from founders.}

\item{mtDNAfreqs}{a data.frame of mtDNA haplotypes ('haplos') in
the first column, and their frequencies ('freqs') in the
second.}
}
\value{
a pop-by-9 character matrix, defined in the \link{makeFounders}
documentation, plus an extra column ('mtHaplo') holding each
animal's mtDNA haplotype.
}
\description{
Adds a column of mtDNA haplotypes to the 'indiv' data of a
completed simulation, given input mtDNA haplotypes and their
frequencies at the time of the founder population. mtDNA haplotypes
are inherited matrilineally (each animal gets its haplotype from
its mother). Mutation is not handled, so mtDNA diversity will
decrease through time via genetic drift, especially in small
populations.
}
\examples{
## set up a population with some founders and their offspring
batchSize = 2
mortRate = 0.5
indiv <- makeFounders(stocks = c(1))
for( y in 1:10) {
    indiv <- altMate( indiv, batchSize = batchSize, type = "flat")
    indiv <- mort( indiv, year = y, type = "flat", mortRate = mortRate)
    indiv <- birthdays(indiv)
}
## make haplotype frequencies
haplos <- paste(expand.grid(LETTERS, LETTERS)[,1],expand.grid(LETTERS, LETTERS)[,2], sep = "")
counts <- ceiling(rexp(length(haplos), rate = 0.2))
freqs <- counts / sum(counts)
mtDNAfreqs <- data.frame(haplos = haplos, freqs = freqs)
## give all animals mtDNA haplotypes consistent with their descent
indiv_haplos <- addmtDNA(indiv, mtDNAfreqs)
}
