% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{altMate}
\alias{altMate}
\title{altMate(): breeding based on mature females, not quota-filling.}
\usage{
altMate(indiv = makeFounders(), batchSize = 0.5,
  fecundityDist = "Poisson", osr = c(0.5, 0.5), year = "-1",
  firstBreed = 0, type = "flat", maxClutch = Inf,
  singlePaternity = TRUE, exhaustFathers = FALSE, maturityCurve,
  maleCurve, femaleCurve)
}
\arguments{
\item{indiv}{A matrix of individuals, e.g., generated in makeFounders() or output from move()}

\item{batchSize}{Numeric. The mean number of offspring produced per mature female if fecundityDist
= "Poisson", or the mean number of offspring per female that gives birth, if
fecundityDist = "truncPoisson". Used for all maturity structures. Note that,
if run within a loop that goes [move -> altMate -> mort -> birthdays],
'produced' is not the same as 'enters age-class 1', as some individuals will
die at age 0.}

\item{fecundityDist}{One of "Poisson" or "truncPoisson". May be expanded to include other
dists in future. ~Bernoulli seems a likely candidate. Defaults to
"Poisson"}

\item{osr}{Numeric vector with length two, c(male, female), giving the sex ratio at birth
(recruitment). Used to assign sexes to new offspring.}

\item{year}{Intended to be used in a simulation loop - this will be the iteration number, and
holds the 'birthyear' value to give to new recruits.}

\item{firstBreed}{Integer variable. The age at first breeding, default zero. The minimum age
at which individuals can breed. Applies to potential mothers and potential
fathers. 'firstBreed', 'maturityCurve', 'maleCurve, and 'femaleCurve' are
all capable of specifying an age at first breeding, and 'firstBreed' takes
precedence.}

\item{type}{The type of maturity-age relationship to simulate. Must be one of "flat",
"age", or "ageSex". If "flat", the probability of parenthood is the same for
all age:sex combinations above firstBreed. If "age", the probability that an
individual is sexually mature is age-specific, set in 'maturityCurve'. If "ageSex",
the probability that an individual is sexually mature is age- and sex-specific,
set for males in 'maleCurve' and for females in 'femaleCurve'.}

\item{maxClutch}{Numeric value giving the maximum clutch / litter / batch / whatever size.
Reduces larger clutches to this size, for each breeding female.}

\item{singlePaternity}{TRUE/FALSE value indicating whether all the offspring produced by
female in a year should have the same father. Default TRUE. If
FALSE, each offspring will have a randomly-drawn father from within
the mother's stock. Note that this can lead to rapid exhaustion of
fathers if exhaustFathers = TRUE.}

\item{exhaustFathers}{TRUE/FALSE value indicating whether fathers should become 'exhausted'
by one breeding attempt. If exhausted, an individual will only mate with
one female, though may father more than one offspring - see
'singlePaternity' and 'batchSize'.}

\item{maturityCurve}{Numeric vector describing the age-specific maturity curve. One
value per age, over all ages from 0:max(indiv[,8]). Used if "type"
= "age". Note that 'firstBreed' can interfere with 'maturityCurve'
by setting maturities to zero for some age classes. Recommended
usage is to set 'firstBreed' to zero whenever 'maturityCurve' is
specified.}

\item{maleCurve}{Numeric vector describing age-specific maturity for males. One value per
age, over all ages from 0:max(indiv[,8]). Used if "type" = "ageSex".
Note that 'firstBreed' can interfere with 'maleCurve' by setting
maturities to zero for some age classes. Recommended usage is to set
'firstBreed' to zero whenever 'maleCurve' is specified.}

\item{femaleCurve}{Numeric vector describing age-specific maturity for females. One value
per age, over all ages from 0:max(indiv[,8]). Used if "type" = "ageSex".
Note that 'firstBreed' can interfere with 'femaleCurve' by setting
maturities to zero for some age classes. Recommended usage is to set
'firstBreed' to zero whenever 'femaleCurve' is specified.}
}
\description{
Returns an individual matrix with added newborns at age 0. This is the second
mate method, where the number of offspring is derived from the number of mature females,
such that each mature female produces a number of offspring specified by a sampling distribution,
and fathers are randomly drawn from all mature males within the mother's stock. Note that in
this mating system, *maturity by age* is specified, rather than *fecundity by age*. A single
probability distribution sets the number of offspring for each female, but the probability
that an individidual female is mature may vary by age. The same maturity by age structure applies
for males. It is possible, in cases where the maturity-by-age slope is shallow, that an individual
may be 'mature' in one breeding season, but then 'not mature' the next season.
Another mate method exists, where the number of newborns is set as a proportion of the population
size, and mating occurs until the required number of offspring are generated (if possible, given
breeding constraints) - see mate().
}
\seealso{
[fishSim::mate()]
}
