% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{archive_dead}
\alias{archive_dead}
\title{Take dead individuals and copy them to an archive}
\usage{
archive_dead(indiv = mort(), archive = make_archive())
}
\arguments{
\item{indiv}{A matrix of individuals, as from \code{\link[=makeFounders]{makeFounders()}}, \code{\link[=move]{move()}},
\code{\link[=mate]{mate()}}, or \code{\link[=mort]{mort()}}.}

\item{archive}{A matrix of individuals, probably from \code{\link[=make_archive]{make_archive()}} or a
previous call of \code{\link[=archive_dead]{archive_dead()}}.}
}
\description{
For larger simulations, the matrix \code{indiv} may grow very large and slow down
the simulation.  In these cases, run-times may be improved by periodically
moving dead individuals into an archive that is read and written less
frequently than \code{indiv}.
}
\examples{
archive <- make_archive()
indiv <- makeFounders()
ages <- min(indiv[,8]):200
ageMort <- 0.1 + (0.2*1/(ages+1))  ## placeholder ageMort with (extreme) negative senescence
stocks <- c(0.3,0.3,0.4)  ## matches defaults in makeFounders
admix.m <- matrix(NA, nrow = length(stocks), ncol = length(stocks))
for (i in 1:nrow(admix.m)) {
     admix.m[i,] <- stocks*stocks[i]
}
indiv <- makeFounders()
for(k in 1:30) {
   indiv <- move(indiv = indiv, moveMat = admix.m)
   indiv <- mate(indiv = indiv, osr = c(0.55,0.45), year = k)
   indiv <- mort(indiv = indiv, type = "age", ageMort = ageMort, year = k)
   indiv <- birthdays(indiv = indiv)
   archive <- archive_dead(indiv = indiv, archive = archive) # archives a copy of dead animals'
                                                             # data
   indiv <- remove_dead(indiv = indiv) # actually removes the dead from 'indiv'.
}
}
\seealso{
\code{\link[=remove_dead]{remove_dead()}} \code{\link[=make_archive]{make_archive()}}
}
