% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{both}
\alias{both}
\title{Unique reals from ordered pairs of integers}
\usage{
both(m1, m2)
}
\arguments{
\item{m1}{a vector of integers}

\item{m2}{another vector of integers}
}
\description{
Return a vector of unique real-valued numbers from a pair of integers
}
