% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{capture}
\alias{capture}
\title{Identify genetic captures/samples in population}
\usage{
capture(
  indiv = makeFounders(),
  n = 1,
  year = "-1",
  fatal = TRUE,
  sex = NULL,
  age = NULL
)
}
\arguments{
\item{indiv}{A matrix of individuals, as from \code{\link[=makeFounders]{makeFounders()}}, \code{\link[=mate]{mate()}}, or
\code{\link[=mort]{mort()}}.}

\item{n}{Number of captures (genetic samples)}

\item{year}{Capture year}

\item{fatal}{Is sampling fatal?}

\item{sex}{Sex specific sampling (either \code{"M"}, \code{"F"} or \code{NULL})}

\item{age}{Integer vector. The age class(es) at which sampling occurs}
}
\description{
Works in a manner similar to \code{\link[=mort]{mort()}}, assigning a year to captured
individuals and killing them if sampling is fatal. Sex specific sampling is
allowed.
}
