% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{check_growthrate}
\alias{check_growthrate}
\title{check_growthrate(): estimate population growth under some mate() and mort() conditions.}
\usage{
check_growthrate(mateType = "flat", mortType = "flat", batchSize,
  firstBreed = 0, maxClutch = Inf, osr = c(0.5, 0.5), maturityCurve,
  femaleCurve, maxAge = Inf, forceY1 = NA, mortRate, ageMort,
  stockMort, ageStockMort)
}
\arguments{
\item{mateType}{the value of 'type' used in the altMate() call. Must be one of 'flat', 'age',
or 'ageSex'. If 'flat', 'batchSize' must be provided. If 'age', 'maturityCurve'
and 'batchSize' must be provided. If 'ageSex', 'femaleCurve' and 'batchSize' must
be provided. Defaults to 'flat'.}

\item{mortType}{the value of 'type' used in the mort() call. Must be one of 'flat', 'age',
'stock', or 'ageStock'. If 'flat', 'mortRate' must be provided. If 'age',
'ageMort' must be provided. If 'stock', 'stockMort' must be provided. If
'ageStock', 'ageStockMort' must be provided. Defaults to 'flat'.}

\item{batchSize}{the value of 'batchSize' used in the altMate() call. Cannot be blank.}

\item{firstBreed}{the value of 'firstBreed' used in the altMate() call. Defaults to 0.}

\item{maxClutch}{the value of 'maxClutch' used in the altMate() call. Defaults to Inf. If non-Inf,
_effective_ batchSize is estimated as the mean of 1000000 draws from the
distribution of batchSize, subsetted to those <= maxAge.}

\item{osr}{the value of 'osr' used in the altMate() call. Female proportion is used as a
multiplier on the fecundities. Defaults to c(0.5, 0.5).}

\item{maturityCurve}{the value of 'maturityCurve' used in the altMate() call. check_growthrates()
only uses female fecundities in its estimates, so femaleCurve is
equivalent to maturityCurve in check_growthrates(), but maturityCurve is
used when mateType is 'age'. If both mortality and maturity are specified
as vectors, they can be of different lengths. If the maturity vector is
shorter, it is 'padded' to the same length as the mortality vector by
repeating the last value in the vector.}

\item{femaleCurve}{the value of 'femaleCurve' used in the altMate() call. check_growthrates()
only uses female fecundities in its estimates, so femaleCurve is
equivalent to maturityCurve in check_growthrates(), but femaleCurve is used
when 'mateType' is 'ageSex'. If both mortality and maturity are specified
as vectors, they can be of different lengths. If the maturity vector is
shorter, it is 'padded' to the same length as the mortality vector by
repeating the last value in the vector.}

\item{maxAge}{the value of 'maxAge' used in the mort() call. Defaults to Inf.}

\item{forceY1}{optionally force first-year mortality to a specific value. Defaults to NA. If non-NA,
should be a numeric value between 0 and 1. If NA, ignored.}

\item{mortRate}{the value of 'mortRate' used in the mort() call}

\item{ageMort}{the value of 'ageMort' used in the mort() call. If both mortality and maturity are
specified as vectors, they can be of different lengths. If the mortality vector is
shorter, it is 'padded' to the same length as the maturity vector by repeating the
last value in the vector.}

\item{stockMort}{the value of 'stockMort' used in the mort() call}

\item{ageStockMort}{the value of 'ageStockMort' used in the mort() call. If both mortality and
maturity are specified as vectors, they can be of different lengths. If the
mortality vector is shorter, it is 'padded' to the same length as the maturity
vector by repeating the last value in the vector.}
}
\description{
In complex models, population trends may not be immediately clear from the settings. Yet
it can be important to know population trends in advance: growing populations take progressively
more computing time, and the relationship dynamics between individuals across generations are
different for growing vs. shrinking populations. check_growthrate() provides an estimate of
the long-term population growth rate under some altMate() and mort() settings, assuming one
altMate() and one mort() per cycle. Estimation is via Leslie matrices.
check_growthrates only functions for some altMate() and mort() structures. Specifically, in
altMate, 'type' must be one of 'flat', 'age', or 'ageSex', and in mort, if 'type' is 'flat'
or 'age', a single growth rate will be returned, but if 'stock' or 'ageStock', one growth
rate will be returned per stock. If 'type' is 'simple' for mort(), the growthrate is forced to
zero, so check_growthrate() does not explicitly handle this case.
Some conditions can cause check_growthrate to fail or provide inaccurate estimates. If mature
females go unmated through lack of available fathers (for instance, if exhaustFathers = TRUE
in mate and N(mature females) > N(mature males) ), the Leslie matrix approach will provide an
over-estimate of the growth rate. In mate(), batchSize is the mean number of offspring per female,
but if maxClutch is also set to a value other than Inf, the *effective* mean number of offspring
per female is estimated by simulation. The mean number of female offspring per female per year
is assumed to be half of the effective mean number of offspring per female unless osr is specified,
in which case the proportion of female offspring is taken from osr.
If your model involves a variation not handled by check_growthrates(), you may find it simplest
to run your simulation for a few generations with a smallish founder population and empirically
estimate the growth rate.
}
\seealso{
[fishSim::PoNG()]
}
