% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{great.grandparents}
\alias{great.grandparents}
\title{look up the great-grandparents of one or more individuals}
\usage{
great.grandparents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from
\code{mort()[,1]}}

\item{indiv}{A matrix of individuals, as from \code{\link[=mort]{mort()}}.}
}
\description{
A convenience wrapper for \code{\link[=parents]{parents()}}, returning the great-grandparents of
one or more individuals. If \code{FFF} is 'father's father's father', MFM is
'mother's father's mother', and so on, great-grandparents are returned in
order \code{FFF}, \code{FFM}, \code{FMF}, \code{FMM}, \code{MFF}, \code{MFM}, \code{MMF}, \code{MMM} for each
specified ID.
}
\seealso{
\code{\link[=parents]{parents()}}
}
