% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{great2.grandparents}
\alias{great2.grandparents}
\title{great2.grandparents(): look up the great-great-grandparents of one or more individuals.}
\usage{
great2.grandparents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from mort()[,1]}

\item{indiv}{A matrix of individuals, as from mort().}
}
\description{
A convenience wrapper for 'parents()', returning the great-great-grandparents of one
or more individuals. If 'FFFF' is 'father's father's father's father', and
'MMFM' is 'mother's mother's father's mother', and so on, great-great-grandparents
are returned in order [FFFF,FFFM,FFMF,FFMM,FMFF,FMFM,FMMF,FMMM,MFFF,MFFM,MFMF,MFMM,
MMFF,MMFM,MMMF,MMMM]
}
\seealso{
[fishSim::parents()]
}
