% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{great3.grandparents}
\alias{great3.grandparents}
\title{Look up the great-great-great-grandparents of one or more individuals.}
\usage{
great3.grandparents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from
\code{mort()[,1]}}

\item{indiv}{A matrix of individuals, as from \code{\link[=mort]{mort()}}.}
}
\description{
A convenience wrapper for \code{\link[=parents]{parents()}}, returning the
great-great-great-grandparents of one or more individuals. If 'FFFFF' is
'father's father's father's father's father', and 'FFMFM' is 'father's
father's mother's father's mother', and so on,
great-great-great-grandparents are returned in order: \code{FFFFF}, \code{FFFFM},
\code{FFFMF}, \code{FFFMM}, \code{FFMFF}, \code{FFMFM}, \code{FFMMF}, \code{FFMMM}, \code{FMFFF}, \code{FMFFM},
\code{FMFMF}, \code{FMFMM}, \code{FMMFF}, \code{FMMFM}, \code{FMMMF}, \code{FMMMM}, \code{MFFFF}, \code{MFFFM},
\code{MFFMF}, \code{MFFMM}, \code{MFMFF}, \code{MFMFM}, \code{MFMMF}, \code{MFMMM}, \code{MMFFF}, \code{MMFFM},
\code{MMFMF}, \code{MMFMM}, \code{MMMFF}, \code{MMMFM}, \code{MMMMF}, \code{MMMMM}.
}
\seealso{
\code{\link[=parents]{parents()}}
}
