% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{make_archive}
\alias{make_archive}
\title{Set up an archive matrix for storing simulation outputs}
\usage{
make_archive()
}
\value{
The function returns an empty 0-by-8 matrix, to which subsets of
\code{indiv} can be attached.
}
\description{
For larger simulations, the matrix \code{indiv} may grow very large and slow down
the simulation.  In these cases, run-times may be improved by periodically
moving dead individuals into an archive that is read and written less
frequently than \code{indiv}.
}
\seealso{
\code{\link[=archive_dead]{archive_dead()}} \code{\link[=remove_dead]{remove_dead()}}
}
