% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{oldmove}
\alias{oldmove}
\title{Markovian movement between breeding stocks}
\usage{
oldmove(indiv = makeFounders(), moveMat)
}
\arguments{
\item{indiv}{Individual matrix, e.g. from \code{makeFounders()}. Can also be a non-founder matrix.}

\item{moveMat}{An s-by-s matrix describing the probability of moving from
each stock to each other stock, with 'from' by row and 'to' by column.
Cannot be blank.}
}
\value{
a pop-by-8 character matrix, defined in the \link{makeFounders}
documentation.
}
\description{
Now superseded by new 'move'. New 'move' will accept the same
inputs for backwards-compatability, but is faster and can handle
age- and/or sex-specific movement patterns.
}
