% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{remove_dead}
\alias{remove_dead}
\title{Remove the dead from a population}
\usage{
remove_dead(indiv = mort())
}
\arguments{
\item{indiv}{A matrix of individuals, as from \code{\link[=mort]{mort()}}.}
}
\description{
For larger simulations, the matrix \code{indiv} may grow very large and slow down
the simulation.  In these cases, run-times may be improved by periodically
moving dead individuals into an archive that is read and written less
frequently than \code{indiv}. See also \code{\link[=archive_dead]{archive_dead()}}.
}
\seealso{
\code{\link[=archive_dead]{archive_dead()}}, \code{\link[=make_archive]{make_archive()}}
}
