% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{rTruncPoisson}
\alias{rTruncPoisson}
\title{Draw from a zero-truncated Poisson dist}
\usage{
rTruncPoisson(n = 1, T = 0.5)
}
\arguments{
\item{n}{The number of values to draw}

\item{T}{The value of lambda for an equivalent non-truncated Poisson}
}
\description{
Returns a vector of draws from a zero-truncated Poisson distribution, with
specified lambda.
}
