% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{findRelatives}
\alias{findRelatives}
\title{findRelatives(): find shared ancestors between pairs of individuals}
\usage{
findRelatives(indiv, sampled)
}
\arguments{
\item{indiv}{A matrix of individuals, as from mort(), but which will need to contain
long strings of parent-offspring relationships, so will most likely come
from a multi-generation simulation.}

\item{sampled}{A character vector containing one or more IDs, e.g., from mort()[,1]}
}
\description{
findRelatives takes a set of 'sampled' individuals and a population simulation output
(i.e., an 'indiv' object, of the kind output by mort() ). It returns a data.frame
for each pair of sampled individuals, with columns:
1) Var1: the first individual's ID
2) Var2: the second individual's ID
3) related: TRUE for each pair that with one or more shared ancestors within seven
   ancestral generations (i.e., great-great-great-great grandparents), otherwise FALSE
4) totalRelatives: numeric indicator of the total number of shared ancestors found
5) OneTwo, OneThree, ThreeFour, etc.: Numeric values, indicating the number of shared
   ancestors by relationship class. For instance, a shared relative in the OneTwo class
   indicates that one member of the pair is the other member's parent, a shared relative
   in the OneThree class indicates that one member is the other's grandparent, and a shared
   in the ThreeFour class indicates that one individual's grandparent is the other's great-
   grandparent. If a pair shares an ancestor in the TwoThree class, they necessarily also
   share two ancestors in the ThreeFour class and four ancestors in the FourFive class,
   and so on. Note that relationship classes are identical by reversal - ThreeFour is the
   same as FourThree, and so only relationship classes with increasing order are presented
   (i.e., ThreeFour and OneFive are in the output, but not ThreeTwo).
}
\seealso{
[fishSim::lookAtPair()]
}
