% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{great.grandparents}
\alias{great.grandparents}
\title{great.grandparents(): look up the great-grandparents of one or more individuals.}
\usage{
great.grandparents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from mort()[,1]}

\item{indiv}{A matrix of individuals, as from mort().}
}
\description{
A convenience wrapper for 'parents()', returning the great-grandparents of one or more
individuals. If FFF is 'father's father's father', MFM is 'mother's father's mother',
and so on, great-grandparents are returned in order [FFF, FFM, FMF, FMM, MFF, MFM,
MMF, MMM] for each specified ID.
}
\seealso{
[fishSim::parents()]
}
