% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{great3.grandparents}
\alias{great3.grandparents}
\title{look up the great-great-great-grandparents of one or more individuals.}
\usage{
great3.grandparents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from mort()[,1]}

\item{indiv}{A matrix of individuals, as from mort().}
}
\description{
A convenience wrapper for 'parents()', returning the great-great-great-grandparents
of one or more individuals. If 'FFFFF' is 'father's father's father's father's father',
and 'FFMFM' is 'father's father's mother's father's mother', and so on,
great-great-great-grandparents are returned in order [FFFFF, FFFFM, FFFMF,
FFFMM, FFMFF, FFMFM, FFMMF, FFMMM, FMFFF, FMFFM, FMFMF, FMFMM,
FMMFF, FMMFM, FMMMF, FMMMM, MFFFF, MFFFM, MFFMF, MFFMM, MFMFF,
MFMFM, MFMMF, MFMMM, MMFFF, MMFFM, MMFMF, MMFMM, MMMFF, MMMFM,
MMMMF, MMMMM]
}
\seealso{
[fishSim::parents()]
}
