% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{make_archive}
\alias{make_archive}
\title{make_archive(): sets up an archive matrix, for storing simulation outputs.}
\usage{
make_archive()
}
\value{
The function returns an empty 0-by-8 matrix, to which subsets of 'indiv' can be
        attached. See archive_dead() and remove_dead().
}
\description{
For larger simulations, the matrix 'indiv' may grow very large and slow down the simulation.
In these cases, run-times may be improved by periodically moving dead individuals into an
archive that is read and written less frequently than 'indiv'.
}
\seealso{
[fishSim::archive_dead(), fishSim::remove_dead()]
}
