% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{makeFounders}
\alias{makeFounders}
\title{Make a founding population}
\usage{
makeFounders(
  pop = 1000,
  osr = c(0.5, 0.5),
  stocks = c(0.3, 0.3, 0.4),
  minAge = 1,
  maxAge = 20,
  survCurv = 0.7^(1:maxAge)/sum(0.7^(1:maxAge))
)
}
\arguments{
\item{pop}{The size of the founder population.}

\item{osr}{A numeric vector describing the sex ratio, \verb{c([male], [female])}.}

\item{stocks}{A numeric vector describing the probability that an individual
is in each stock.}

\item{maxAge}{Numeric. The max age to which an animal may survive.}

\item{survCurv}{Numeric vector. Describes the probability within the founder
cohort of belonging to each age-class for \verb{age=-classes 1:maxAge}. Cannot be
blank. Will be treated as probability weights, rather than probabilities, if
the vector does not sum to 1.}
}
\description{
Returns a population-size-by-8 \code{data.frame}, with each row being an
individual in the founder population.
\itemize{
\item \verb{[,1]} is a unique (uuid) identifier for each animal.
\item \verb{[,2]} is sex, values \code{"M"} or \code{"F"}.
\item \verb{[,3]} is \code{"founder"} for all animals.
\item \verb{[,4]} is \code{"founder"} for all animals.
\item \verb{[,5]} is the animal's birth year. Implicitly assumes that \code{makeFounders}
occurs at the very start of year 1, just after the \code{birthdays} step of year 0.
\item \verb{[,6]} is \code{NA} for all animals. Holds the death year in most cases.
\item \verb{[,7]} is the stock membership for each animal.
\item \verb{[,8]} is the age of each animal (in 'breeding seasons') at the beginning
of year 1, given that birthdays occur at the very end.
\item \verb{[,9]} is \code{NA} for all animals
}
}
\details{
\code{makeFounders()} will throw a warning if \code{osr}, \code{stocks}, or \code{survCurv} do
not sum to 1. It is not strictly necessary that they sum to 1
(proportionality within each class is sufficient), but error-checking and
readability is easiest if they do sum to 1.
}
\seealso{
\code{\link[=make_archive]{make_archive()}}
}
