% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{mate}
\alias{mate}
\title{mate(): find male-female pairs within the same stock, and generate batches of offspring
until a recruitment quota is filled.}
\usage{
mate(indiv = makeFounders(), fecundity = 0.2, batchSize = 0.5,
  osr = c(0.5, 0.5), year = "-1", firstBreed = 0)
}
\arguments{
\item{indiv}{A matrix of individuals, e.g., generated in makeFounders() or output from move()}

\item{fecundity}{mean number of recruits to generate, as a proportion of the number
of animals in 'indiv'. nrow(indiv)*fecundity is the annual turnover -
i.e., the number of animals killed in 'mortality' will equal
nrow(indiv)*fecundity in order to keep the population size constant.}

\item{batchSize}{The mean number of offspring produced per pairing. Follows ~Poisson.}

\item{osr}{The sex ratio at birth (recruitment). Used to assign sexes to new offspring.}

\item{year}{Intended to be used in a simulation loop - this will be the iteration number, and
holds the 'birthyear' value to give to new recruits.}

\item{firstBreed}{The age at first breeding, default zero. The minimum age at which individuals
can breed. Applies to both potential mothers and potential fathers.}
}
\description{
Returns an individual matrix with added newborns at age 0.
}
