% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{mate}
\alias{mate}
\title{mate(): find male-female pairs within the same stock, and generate batches of offspring
until a recruitment quota is filled.}
\usage{
mate(indiv = makeFounders(), fecundity = 0.2, batchSize = 0.5,
  osr = c(0.5, 0.5), year = "-1", firstBreed = 1, type = "flat",
  maxClutch = Inf, exhaustMothers = FALSE, exhaustFathers = FALSE,
  fecundityCurve, maleCurve, femaleCurve)
}
\arguments{
\item{indiv}{A matrix of individuals, e.g., generated in makeFounders() or output from move()}

\item{fecundity}{Numeric variable. mean number of recruits to generate, as a proportion of
the number of animals in 'indiv'. nrow(indiv)*fecundity is the annual
turnover - i.e., the number of animals killed in 'mortality' will equal
nrow(indiv)*fecundity in order to keep the population size constant.}

\item{batchSize}{Numeric. The mean number of offspring produced per pairing. Follows ~Poisson.
Used iff type = "flat". Note that, if run within a loop that
goes [move -> mate -> mort -> birthdays], 'produced' is not the same as 'enters
age-class 1', as some mortality may occur at age 0.}

\item{osr}{Numeric vector with length two, c(male, female), giving the sex ratio at birth
(recruitment). Used to assign sexes to new offspring.}

\item{year}{Intended to be used in a simulation loop - this will be the iteration number, and
holds the 'birthyear' value to give to new recruits.}

\item{firstBreed}{Integer variable. The age at first breeding, default 1. The minimum age
at which individuals can breed. Applies to potential mothers and potential
fathers. Both firstBreed and 'fecundityCurve', 'maleCurve, and 'femaleCurve' are
capable of specifying an age at first breeding, and firstBreed takes precedence.}

\item{type}{The type of fecundity-age relationship to simulate. Must be one of "flat",
"age", or "ageSex". If "flat", offspring batch sizes will be the same for
all age:sex combinations. If "age", the number of offspring for each pairing
is a function of 'fecundityCurve', with the less-fecund parent determining the
batch size. If "ageSex", the number of offspring for each pairing is a function
of 'maleCurve' and 'femaleCurve', with the less-fecund parent determining the
batch size.}

\item{maxClutch}{Numeric value giving the maximum clutch / litter / batch / whatever size.
Reduces larger clutches to this size, within each pairing.}

\item{exhaustMothers}{TRUE/FALSE value indicating whether mothers should become 'exhausted'
by one breeding attempt. If exhausted, an individual will not mate
again in this mate() call.}

\item{exhaustFathers}{TRUE/FALSE value indicating whether fathers should become 'exhausted'
by one breeding attempt. If exhausted, an individual will not mate
again in this mate() call.}

\item{fecundityCurve}{Numeric vector describing the age-specific fecundity curve. One
value per age, over all ages from 0:max(indiv[,8]). Used if "type"
= "age". Note that 'firstBreed' can interfere with 'fecundityCurve'
by setting fecundities to zero for some age classes. Recommended
usage is to set 'firstBreed' to zero whenever 'fecundityCurve' is
specified.}

\item{maleCurve}{Numeric vector describing age-specific fecundity for males. One value per
age, over all ages from 0:max(indiv[,8]). Used if "type" = "ageSex".
Note that 'firstBreed' can interfere with 'maleCurve' by setting
fecundities to zero for some age classes. Recommended usage is to set
'firstBreed' to zero whenever 'maleCurve' is specified.}

\item{femaleCurve}{Numeric vector describing age-specific fecundity for females. One value
per age, over all ages from 0:max(indiv[,8]). Used if "type" = "ageSex".
Note that 'firstBreed' can interfere with 'femaleCurve' by setting
fecundities to zero for some age classes. Recommended usage is to set
'firstBreed' to zero whenever 'femaleCurve' is specified.}
}
\description{
Returns an individual matrix with added newborns at age 0. This is the first
mate method, where the number of newborns is a set as a proportion of the adult population,
and matings happen between individuals that are older than the age at first-breeding
(optionally with breeding success-rates set by the age of the less-mature parent, etc.,) until
that 'quota' or newborns has been met, or all potential parents have reached breeding exhaustion.
A second mate method exists, where the number of newborns is derived from the fecundities of all
breeding females in the population - see altMate().
}
\seealso{
[fishSim::altMate()]
}
