% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{move}
\alias{move}
\title{markovian movement between breeding stocks}
\usage{
move(indiv = makeFounders(), moveMat)
}
\arguments{
\item{indiv}{Individual matrix, e.g. from makeFounders(). Can also be a non-founder matrix.}

\item{moveMat}{An s-by-s matrix describing the probability of moving from each stock to
each other stock, with 'from' by row and 'to' by column. Cannot be blank.}
}
\description{
returns a pop-by-8 character matrix, defined in the makeFounders documentation.
}
