% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{parents}
\alias{parents}
\title{parents(): look up the parents of one or more individuals.}
\usage{
parents(ID, indiv)
}
\arguments{
\item{ID}{A character vector containing one or more IDs, e.g., from mort()[,1]}

\item{indiv}{A matrix of individuals, as from mort().}
}
\description{
The parents() function searches an 'indiv' matrix for individuals with a given ID,
and returns the IDs of its parents. If one or more of the parents were founders, it
returns 'founder' instead of a ID for founder parents, and if the individual's ID
is 'founder', it returns 'founder' as the ID for both parents. Parents are returned
in order [father, mother].
If more than one ID is given, parents are returned in order of ID: [ID1 father,
ID1 mother, ID2 father, ID2 mother ... ]
}
\examples{
indiv <- makeFounders()
for(k in 1:30) {
  indiv <- mate(indiv = indiv, year = k)
  indiv <- mort(indiv = indiv, year = k)
  indiv <- birthdays(indiv = indiv)
}  ## set up a population
ID <- indiv[nrow(indiv), 1] ## an individual from the youngest generation
parents(ID, indiv) ## that individual's parents, as [father, mother]
parents(parents(ID, indiv), indiv) ## that individual's grandparents, as
## [pat. grandfather, pat. grandmother, mat. grandfather, mat. grandmother]
}
\seealso{
[fishSim::grandparents(), fishSim::great.grandparents(),
fishSim::great4.grandparents()]
}
