% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{sexSwitch}
\alias{sexSwitch}
\title{sexSwitch(): induces sex-switching in a subset of the population.}
\usage{
sexSwitch(indiv = makeFounders(), direction = "both", prob = 1e-04)
}
\arguments{
\item{indiv}{A matrix of individuals, as from makeFounders(), mate(), or mort().}

\item{direction}{One of "MF", "FM", or "both", indicating male-to-female, female-to-male,
or both directions.}

\item{prob}{A numeric switching probability.}
}
\description{
Given an individual-data matrix, this function stochastically switches the sex of individuals
with a set probability. Can be one-directional (i.e., only male-to-female or only female-to-male
switches), or bidirectional.
}
